[fix] Stable Audio: match official's fp16 dtype + enable FlashAttention
The official `stable_audio_tools` ships with `model_half=True` by default
(`run_gradio.py:31`), loading the DiT, VAE, and T5 in fp16. Our port
defaulted to fp32 everywhere, which (a) diverged from upstream and (b)
disqualified FlashAttention since FA-2/3/4 only accept fp16/bf16.

Changes:

- `StableAudioT2AConfig.precision` and `vae_precision` → `"fp16"`.
- `OobleckVAEConfig.pretrained_dtype` → `"float16"` (matches the
  `pretransform.model_half` path in stable_audio_tools).
- `T5Conditioner` loads T5 directly in fp16 + wraps the forward in
  `torch.autocast(fp16)` (mirrors `conditioners.py:334` and `:362`).
- Pipeline `load_modules` now also casts the conditioner to
  `torch_dtype` and calls `set_mixed_precision_policy(fp16)` before
  attention layers are constructed — without this,
  `fastvideo.attention.layer` reads `torch.get_default_dtype()` (fp32)
  and rejects FA. The standard FSDP loader does this in
  `fsdp_load.py:89`; we bypass that loader so we set it ourselves.
- `StableAudioDenoisingStage` casts the noise + conditioning tensors to
  the DiT dtype just before sampling (mirrors
  `stable_audio_tools/inference/generation.py:185-187`), and casts the
  RePaint inpaint mask/ref to model dtype too (was fp32, would
  dtype-mismatch the fp16 sampled latent).

+++ fastvideo/configs/models/vaes/oobleck.py
@@ -62,4 +62,7 @@ class OobleckVAEConfig(VAEConfig):
     # https://huggingface.co/stabilityai/stable-audio-open-1.0.
     pretrained_path: str = "stabilityai/stable-audio-open-1.0"
     pretrained_subfolder: str = "vae"
-    pretrained_dtype: str = "float32"
+    # Match official `stable_audio_tools`: VAE runs in fp16 (the
+    # `pretransform.model_half` path in
+    # `stable_audio_tools/models/pretransforms.py`).
+    pretrained_dtype: str = "float16"
